cart functionality

+++ src/app/modules/cart/cart.controller.ts
@@ -3,10 +3,22 @@ import catchAsync from '../../utils/catchAsync';
 import sendResponse from '../../utils/sendResponse';
 import httpStatus from 'http-status';
 import * as cartService from './cart.service';
+import { CartValidation } from './cart.validation';
 
 const addToCart = catchAsync(async (req: Request, res: Response) => {
-  const userId = req.user.id;
+  const userId = req.user?.id;
   const { productId, quantity } = req.body;
+
+  if (!userId) {
+    return sendResponse(res, {
+      statusCode: httpStatus.UNAUTHORIZED,
+      success: false,
+      message: 'Unauthorized',
+      data: null,
+    });
+  }
+
+  CartValidation.addToCartSchema.parse(req.body);
 
   const cart = await cartService.addToCart(userId, productId, quantity);
 
@@ -19,9 +31,27 @@ const addToCart = catchAsync(async (req: Request, res: Response) => {
 });
 
 const getCart = catchAsync(async (req: Request, res: Response) => {
-  const userId = req.user.id;
+  const userId = req.user?.id;
+
+  if (!userId) {
+    return sendResponse(res, {
+      statusCode: httpStatus.UNAUTHORIZED,
+      success: false,
+      message: 'Unauthorized',
+      data: null,
+    });
+  }
 
   const cart = await cartService.getCart(userId);
+
+  if (!cart) {
+    return sendResponse(res, {
+      statusCode: httpStatus.NOT_FOUND,
+      success: false,
+      message: 'Cart is empty',
+      data: null,
+    });
+  }
 
   sendResponse(res, {
     statusCode: httpStatus.OK,
@@ -32,8 +62,19 @@ const getCart = catchAsync(async (req: Request, res: Response) => {
 });
 
 const updateCart = catchAsync(async (req: Request, res: Response) => {
-  const userId = req.user.id;
+  const userId = req.user?.id;
   const { productId, quantity } = req.body;
+
+  if (!userId) {
+    return sendResponse(res, {
+      statusCode: httpStatus.UNAUTHORIZED,
+      success: false,
+      message: 'Unauthorized',
+      data: null,
+    });
+  }
+
+  CartValidation.updateCartSchema.parse(req.body);
 
   const cart = await cartService.updateCart(userId, productId, quantity);
 
@@ -46,8 +87,19 @@ const updateCart = catchAsync(async (req: Request, res: Response) => {
 });
 
 const removeFromCart = catchAsync(async (req: Request, res: Response) => {
-  const userId = req.user.id;
+  const userId = req.user?.id;
   const { productId } = req.body;
+
+  if (!userId) {
+    return sendResponse(res, {
+      statusCode: httpStatus.UNAUTHORIZED,
+      success: false,
+      message: 'Unauthorized',
+      data: null,
+    });
+  }
+
+  CartValidation.removeFromCartSchema.parse(req.body);
 
   const cart = await cartService.removeFromCart(userId, productId);
